Move state and error handling into the page itself.
State: this state is specific to the page itself only anyway.

Error handling: same as above and considering it's calling toast.error(), having it inside a svelte file makes more sense.
