+++ src/frontend/src/routes/(new-styling)/new-authorize/components/CreatePasskey.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import Button from "$lib/components/UI/Button.svelte";
   import { fly } from "svelte/transition";
-  import { type State } from "../state";
+  import { type State } from "../+page.svelte";
   import { onMount } from "svelte";
 
   type Props = Extract<State, { state: "createPasskey" }>;
