--- src/frontend/src/routes/(new-styling)/new-authorize/state.ts
@@ -1,23 +1,0 @@
-export type State =
-  | { state: "loading" }
-  | { state: "continueAs"; number: bigint; name?: string }
-  | {
-      state: "pickAuthenticationMethod";
-    }
-  | {
-      state: "connectOrCreatePasskey";
-      connect: () => void;
-      create: () => void;
-    }
-  | {
-      state: "createPasskey";
-      create: (name: string) => void;
-      cancel: () => void;
-    }
-  | {
-      state: "solveCaptcha";
-      image: string;
-      attempt: number;
-      solve: (solution: string) => void;
-      cancel: () => void;
-    };
